multiple stages, more hovery control, better enemy AI
